+++ src/Entity.store.test.ts
@@ -30,7 +30,7 @@ describe("EntityStore", function () {
 
         await afterAllStoreEvents(es);
 
-        let changes = collect(es.unbounded().selectById("1"));
+        let changes = collect(es.selectById("1").unbounded());
 
         changes.assert(
             new MyEnt(1, "a")
@@ -39,7 +39,7 @@ describe("EntityStore", function () {
 
     it("selectById() emits undefined until ID exists", async function () {
         let es = new EntityStore("es1", MyEnt, "id");
-        let changes = collect(es.unbounded().selectById("1"));
+        let changes = collect(es.selectById("1").unbounded());
         es.add(new MyEnt(1, "a"));
 
         await afterAllStoreEvents(es);
@@ -52,7 +52,7 @@ describe("EntityStore", function () {
 
     it("selectAll()", async function () {
         let es = new EntityStore("es1", MyEnt, "id");
-        let changes = collect(es.unbounded().selectAll());
+        let changes = collect(es.selectAll().unbounded());
 
         es.add(new MyEnt(1, "a"));
         es.add(new MyEnt(2, "b"));